Remove myself from the code-of-conduct list
I think conferences having a code-of-conduct is a good thing and that it is to be encouraged. My pledge to not attend conferences without a code-of-conduct was, however, hasty on my part. I'm already planning to attend Clojure/conj, which does not (yet) have a published code-of-conduct. I will suggest that they publish one, but understand that publishing one now for a mid-November conference doesn't quite make sense.